Move OpenSSL 3 implementation to separate file
