+++ src/soter/openssl/soter_ec_key.c
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
+
 #include "soter/soter_ec_key.h"
 
 #include <string.h>
@@ -21,14 +23,9 @@
 #include <openssl/bn.h>
 #include <openssl/ec.h>
 #include <openssl/evp.h>
-#ifdef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
-#include <openssl/core_names.h>
-#include <openssl/param_build.h>
-#endif
 
 #include "soter/soter_portable_endian.h"
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
 static bool is_curve_supported(int curve)
 {
     switch (curve) {
@@ -40,23 +37,8 @@ static bool is_curve_supported(int curve)
         return false;
     }
 }
-#else
-static bool is_curve_supported_2(const char* curve)
-{
-    if (strcmp(curve, "prime256v1" /* P-256 */) == 0) {
-        return true;
-    } else if (strcmp(curve, "secp384r1" /* P-384 */) == 0) {
-        return true;
-    } else if (strcmp(curve, "secp521r1" /* P-521 */) == 0) {
-        return true;
-    } else {
-        return false;
-    }
-}
-#endif
 
 /* Input size directly since public key type structures may be aligned to word boundary */
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
 static size_t ec_pub_key_size(int curve, bool compressed)
 {
     switch (curve) {
@@ -73,25 +55,7 @@ static size_t ec_pub_key_size(int curve, bool compressed)
         return 0;
     }
 }
-#else
-static size_t ec_pub_key_size_2(const char* curve, bool compressed)
-{
-    if (strcmp(curve, "prime256v1" /* P-256 */) == 0) {
-        return sizeof(soter_container_hdr_t)
-               + (compressed ? EC_PUB_SIZE(256) : EC_PUB_UNCOMPRESSED_SIZE(256));
-    } else if (strcmp(curve, "secp384r1" /* P-384 */) == 0) {
-        return sizeof(soter_container_hdr_t)
-               + (compressed ? EC_PUB_SIZE(384) : EC_PUB_UNCOMPRESSED_SIZE(384));
-    } else if (strcmp(curve, "secp521r1" /* P-521 */) == 0) {
-        return sizeof(soter_container_hdr_t)
-               + (compressed ? EC_PUB_SIZE(521) : EC_PUB_UNCOMPRESSED_SIZE(521));
-    } else {
-        return 0;
-    }
-}
-#endif
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
 static size_t ec_priv_key_size(int curve)
 {
     switch (curve) {
@@ -105,22 +69,7 @@ static size_t ec_priv_key_size(int curve)
         return 0;
     }
 }
-#else
-static size_t ec_priv_key_size_2(const char* curve)
-{
-    if (strcmp(curve, "prime256v1" /* P-256 */) == 0) {
-        return sizeof(soter_ec_priv_key_256_t);
-    } else if (strcmp(curve, "secp384r1" /* P-384 */) == 0) {
-        return sizeof(soter_ec_priv_key_384_t);
-    } else if (strcmp(curve, "secp521r1" /* P-521 */) == 0) {
-        return sizeof(soter_ec_priv_key_521_t);
-    } else {
-        return 0;
-    }
-}
-#endif
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
 static char* ec_pub_key_tag(int curve)
 {
     switch (curve) {
@@ -134,22 +83,7 @@ static char* ec_pub_key_tag(int curve)
         return NULL;
     }
 }
-#else
-static char* ec_pub_key_tag_2(const char* curve)
-{
-    if (strcmp(curve, "prime256v1" /* P-256 */) == 0) {
-        return EC_PUB_KEY_TAG(256);
-    } else if (strcmp(curve, "secp384r1" /* P-384 */) == 0) {
-        return EC_PUB_KEY_TAG(384);
-    } else if (strcmp(curve, "secp521r1" /* P-521 */) == 0) {
-        return EC_PUB_KEY_TAG(521);
-    } else {
-        return NULL;
-    }
-}
-#endif
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
 static char* ec_priv_key_tag(int curve)
 {
     switch (curve) {
@@ -163,20 +97,6 @@ static char* ec_priv_key_tag(int curve)
         return NULL;
     }
 }
-#else
-static char* ec_priv_key_tag(const char* curve)
-{
-    if (strcmp(curve, "prime256v1" /* P-256 */) == 0) {
-        return EC_PRIV_KEY_TAG(256);
-    } else if (strcmp(curve, "secp384r1" /* P-384 */) == 0) {
-        return EC_PRIV_KEY_TAG(384);
-    } else if (strcmp(curve, "secp521r1" /* P-521 */) == 0) {
-        return EC_PRIV_KEY_TAG(521);
-    } else {
-        return NULL;
-    }
-}
-#endif
 
 static size_t bn_encode(const BIGNUM* bn, uint8_t* buffer, size_t length)
 {
@@ -196,22 +116,15 @@ soter_status_t soter_engine_specific_to_ec_pub_key(const soter_engine_specific_e
     EVP_PKEY* pkey = (EVP_PKEY*)engine_key;
     soter_status_t res;
     size_t output_length;
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     EC_KEY* ec;
     const EC_GROUP* group;
     const EC_POINT* Q;
     int curve;
-#else
-    char group_str[16];
-    const char* param_name;
-    size_t serialized_len;
-#endif
 
     if ((!key_length) || (EVP_PKEY_EC != EVP_PKEY_id(pkey))) {
         return SOTER_INVALID_PARAMETER;
     }
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     ec = EVP_PKEY_get1_EC_KEY((EVP_PKEY*)pkey);
     if (NULL == ec) {
         return SOTER_FAIL;
@@ -232,24 +145,6 @@ soter_status_t soter_engine_specific_to_ec_pub_key(const soter_engine_specific_e
     }
 
     output_length = ec_pub_key_size(curve, compressed);
-#else
-    // Curve identifier as string ("prime256v1", "secp384r1", "secp521r1" etc)
-    if (!EVP_PKEY_get_utf8_string_param(pkey,
-                                        OSSL_PKEY_PARAM_GROUP_NAME /* "group" */,
-                                        group_str,
-                                        sizeof(group_str),
-                                        NULL)) {
-        res = SOTER_INVALID_PARAMETER;
-        goto err;
-    }
-
-    if (!is_curve_supported_2(group_str)) {
-        res = SOTER_INVALID_PARAMETER;
-        goto err;
-    }
-
-    output_length = ec_pub_key_size_2(group_str, compressed);
-#endif
     if ((!key) || (output_length > *key_length)) {
         *key_length = output_length;
         res = SOTER_BUFFER_TOO_SMALL;
@@ -258,7 +153,6 @@ soter_status_t soter_engine_specific_to_ec_pub_key(const soter_engine_specific_e
 
     *key_length = output_length;
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     Q = EC_KEY_get0_public_key(ec);
     if (NULL == Q) {
         res = SOTER_INVALID_PARAMETER;
@@ -275,48 +169,16 @@ soter_status_t soter_engine_specific_to_ec_pub_key(const soter_engine_specific_e
         res = SOTER_FAIL;
         goto err;
     }
-#else
-    if (compressed) {
-        param_name = OSSL_PKEY_PARAM_PUB_KEY /* "pub" */;
 
-        if (!EVP_PKEY_set_utf8_string_param(pkey, "point-format", "compressed")) {
-            res = SOTER_FAIL;
-            goto err;
-        }
-    } else {
-        param_name = OSSL_PKEY_PARAM_ENCODED_PUBLIC_KEY /* "encoded-pub-key" */;
-    }
-
-    if (!EVP_PKEY_get_octet_string_param(pkey,
-                                         param_name,
-                                         (unsigned char*)(key + 1),
-                                         output_length - sizeof(soter_container_hdr_t),
-                                         &serialized_len)) {
-        res = SOTER_INVALID_PARAMETER;
-        goto err;
-    }
-
-    if (serialized_len != output_length - sizeof(soter_container_hdr_t)) {
-        res = SOTER_INVALID_PARAMETER;
-        goto err;
-    }
-#endif
-
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     memcpy(key->tag, ec_pub_key_tag(curve), SOTER_CONTAINER_TAG_LENGTH);
-#else
-    memcpy(key->tag, ec_pub_key_tag_2(group_str), SOTER_CONTAINER_TAG_LENGTH);
-#endif
     key->size = htobe32(output_length);
     soter_update_container_checksum(key);
     *key_length = output_length;
     res = SOTER_SUCCESS;
 
 err:
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     /* Free extra reference on EC_KEY object provided by EVP_PKEY_get1_EC_KEY */
     EC_KEY_free(ec);
-#endif
 
     return res;
 }
@@ -329,21 +191,15 @@ soter_status_t soter_engine_specific_to_ec_priv_key(const soter_engine_specific_
     const bool compressed = true;
     soter_status_t res;
     size_t output_length;
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     EC_KEY* ec;
     const EC_GROUP* group;
     const BIGNUM* d;
     int curve;
-#else
-    char group_str[16];
-    BIGNUM* d = NULL;
-#endif
 
     if ((!key_length) || (EVP_PKEY_EC != EVP_PKEY_id(pkey))) {
         return SOTER_INVALID_PARAMETER;
     }
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     ec = EVP_PKEY_get1_EC_KEY((EVP_PKEY*)pkey);
     if (NULL == ec) {
         return SOTER_FAIL;
@@ -379,40 +235,6 @@ soter_status_t soter_engine_specific_to_ec_priv_key(const soter_engine_specific_
         res = SOTER_INVALID_PARAMETER;
         goto err;
     }
-#else
-    // Curve identifier as string ("prime256v1", "secp384r1", "secp521r1" etc)
-    if (!EVP_PKEY_get_utf8_string_param(pkey,
-                                        OSSL_PKEY_PARAM_GROUP_NAME /* "group" */,
-                                        group_str,
-                                        sizeof(group_str),
-                                        NULL)) {
-        res = SOTER_INVALID_PARAMETER;
-        goto err;
-    }
-
-    if (!is_curve_supported_2(group_str)) {
-        res = SOTER_INVALID_PARAMETER;
-        goto err;
-    }
-
-    /*
-     * Note that we use a buffer suitable for a public key to store a private
-     * key. This was a historical mistake, now preserved for compatibility.
-     */
-    output_length = ec_pub_key_size_2(group_str, compressed);
-    if ((!key) || (output_length > *key_length)) {
-        *key_length = output_length;
-        res = SOTER_BUFFER_TOO_SMALL;
-        goto err;
-    }
-
-    *key_length = output_length;
-
-    if (!EVP_PKEY_get_bn_param(pkey, OSSL_PKEY_PARAM_PRIV_KEY /* "priv" */, &d)) {
-        res = SOTER_INVALID_PARAMETER;
-        goto err;
-    }
-#endif
 
     if ((output_length - sizeof(soter_container_hdr_t))
         != bn_encode(d, (unsigned char*)(key + 1), output_length - sizeof(soter_container_hdr_t))) {
@@ -420,21 +242,15 @@ soter_status_t soter_engine_specific_to_ec_priv_key(const soter_engine_specific_
         goto err;
     }
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     memcpy(key->tag, ec_priv_key_tag(curve), SOTER_CONTAINER_TAG_LENGTH);
-#else
-    memcpy(key->tag, ec_priv_key_tag(group_str), SOTER_CONTAINER_TAG_LENGTH);
-#endif
     key->size = htobe32(output_length);
     soter_update_container_checksum(key);
     *key_length = output_length;
     res = SOTER_SUCCESS;
 
 err:
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     /* Free extra reference on EC_KEY object provided by EVP_PKEY_get1_EC_KEY */
     EC_KEY_free(ec);
-#endif
 
     return res;
 }
@@ -443,18 +259,12 @@ soter_status_t soter_ec_pub_key_to_engine_specific(const soter_container_hdr_t* 
                                                    size_t key_length,
                                                    soter_engine_specific_ec_key_t** engine_key)
 {
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     int curve;
     EC_KEY* ec = NULL;
     const EC_GROUP* group;
     EC_POINT* Q = NULL;
     EVP_PKEY* pkey = (EVP_PKEY*)(*engine_key);
     const bool compressed = true;
-#else
-    const char* curve_str;
-    OSSL_PARAM params[3] = {[2] = OSSL_PARAM_END};
-    EVP_PKEY_CTX* ctx = NULL;
-#endif
     soter_status_t res;
 
     if ((!key) || (key_length < sizeof(soter_container_hdr_t))) {
@@ -474,7 +284,6 @@ soter_status_t soter_ec_pub_key_to_engine_specific(const soter_container_hdr_t* 
         return SOTER_DATA_CORRUPT;
     }
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     switch (key->tag[3]) {
     case EC_SIZE_TAG_256:
         curve = NID_X9_62_prime256v1;
@@ -535,51 +344,9 @@ soter_status_t soter_ec_pub_key_to_engine_specific(const soter_container_hdr_t* 
     } else {
         res = SOTER_FAIL;
     }
-#else
-    switch (key->tag[3]) {
-    case EC_SIZE_TAG_256:
-        curve_str = "prime256v1";
-        break;
-    case EC_SIZE_TAG_384:
-        curve_str = "secp384r1";
-        break;
-    case EC_SIZE_TAG_521:
-        curve_str = "secp521r1";
-        break;
-    default:
-        return SOTER_INVALID_PARAMETER;
-    }
-
-    params[0] = OSSL_PARAM_construct_utf8_string(OSSL_PKEY_PARAM_GROUP_NAME /* "group" */,
-                                                 (char*)curve_str,
-                                                 strlen(curve_str));
-
-    params[1] = OSSL_PARAM_construct_octet_string(OSSL_PKEY_PARAM_PUB_KEY /* "pub" */,
-                                                  (unsigned char*)(key + 1),
-                                                  (int)(key_length - sizeof(soter_container_hdr_t)));
-
-    ctx = EVP_PKEY_CTX_new_from_name(NULL, "EC", NULL);
-    if (ctx == NULL) {
-        res = SOTER_FAIL;
-        goto err;
-    }
-
-    if (!EVP_PKEY_fromdata_init(ctx)) {
-        res = SOTER_FAIL;
-        goto err;
-    }
-
-    if (!EVP_PKEY_fromdata(ctx, (EVP_PKEY**)engine_key, EVP_PKEY_PUBLIC_KEY, params)) {
-        res = SOTER_FAIL;
-        goto err;
-    }
-
-    res = SOTER_SUCCESS;
-#endif
 
 err:
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     if (Q) {
         EC_POINT_free(Q);
     }
@@ -587,9 +354,6 @@ err:
     if (ec) {
         EC_KEY_free(ec);
     }
-#else
-    EVP_PKEY_CTX_free(ctx);
-#endif
 
     return res;
 }
@@ -598,18 +362,11 @@ soter_status_t soter_ec_priv_key_to_engine_specific(const soter_container_hdr_t*
                                                     size_t key_length,
                                                     soter_engine_specific_ec_key_t** engine_key)
 {
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     int curve;
     EC_KEY* ec = NULL;
     const EC_GROUP* group;
-    EVP_PKEY* pkey = (EVP_PKEY*)(*engine_key);
-#else
-    char curve_str[16];
-    OSSL_PARAM* params = NULL;
-    OSSL_PARAM_BLD *bld = NULL;
-    EVP_PKEY_CTX* ctx = NULL;
-#endif
     BIGNUM* d = NULL;
+    EVP_PKEY* pkey = (EVP_PKEY*)(*engine_key);
     soter_status_t res;
 
     if (key_length != be32toh(key->size)) {
@@ -625,7 +382,6 @@ soter_status_t soter_ec_priv_key_to_engine_specific(const soter_container_hdr_t*
         return SOTER_DATA_CORRUPT;
     }
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     switch (key->tag[3]) {
     case EC_SIZE_TAG_256:
         curve = NID_X9_62_prime256v1;
@@ -675,92 +431,9 @@ soter_status_t soter_ec_priv_key_to_engine_specific(const soter_container_hdr_t*
     } else {
         res = SOTER_FAIL;
     }
-#else
-    switch (key->tag[3]) {
-    case EC_SIZE_TAG_256:
-        strcpy(curve_str, "prime256v1");
-        break;
-    case EC_SIZE_TAG_384:
-        strcpy(curve_str, "secp384r1");
-        break;
-    case EC_SIZE_TAG_521:
-        strcpy(curve_str, "secp521r1");
-        break;
-    default:
-        return SOTER_INVALID_PARAMETER;
-    }
-
-    if (key_length < ec_priv_key_size_2(curve_str)) {
-        return SOTER_INVALID_PARAMETER;
-    }
-
-    bld = OSSL_PARAM_BLD_new();
-    if (bld == NULL) {
-        res = SOTER_NO_MEMORY;
-        goto err;
-    }
-
-    if (!OSSL_PARAM_BLD_push_utf8_string(bld,
-                                         OSSL_PKEY_PARAM_GROUP_NAME /* "group" */,
-                                         curve_str,
-                                         strlen(curve_str))) {
-        res = SOTER_FAIL;
-        goto err;
-    }
-
-    d = BN_bin2bn((const unsigned char*)(key + 1), (int)(key_length - sizeof(soter_container_hdr_t)), NULL);
-    if (d == NULL) {
-        res = SOTER_NO_MEMORY;
-        goto err;
-    }
-
-    // Unfortunately, using on stack `OSSL_PARAM params[3]`
-    // like in soter_ec_pub_key_to_engine_specific() and setting
-    // params[1] = OSSL_PARAM_construct_BN(OSSL_PKEY_PARAM_PRIV_KEY /* "priv" */,
-    //                                     (unsigned char*)(key + 1) + 1,
-    //                                     (int)(key_length - sizeof(soter_container_hdr_t) - 1)))
-    // won't work properly. This is because private key number (32 bytes, not counting leading zero)
-    // is stored in one byte order and EVP_PKEY_fromdata() will read it in different byte order. As
-    // a result, `priv` property of generated `pkey` will have different value, with reversed bytes
-    // order. One working solution was to do like this, deserialize value into `BIGINT*`, then use
-    // param builder (OSSL_PARAM_BLD*) and OSSL_PARAM_BLD_push_BN() to essentially write that bigint
-    // into temporary buffer but in different byte order, so EVP_PKEY_fromdata() will work as
-    // intended. Of course, we could also simply create handwritten function for bytes reversing.
-    // And better solution could exist hidden somewhere in OpenSSL API.
-
-    if (!OSSL_PARAM_BLD_push_BN(bld, OSSL_PKEY_PARAM_PRIV_KEY /* "priv" */, d)) {
-        res = SOTER_FAIL;
-        goto err;
-    }
-
-    params = OSSL_PARAM_BLD_to_param(bld);
-    if (params == NULL) {
-        res = SOTER_NO_MEMORY;
-        goto err;
-    }
-
-    ctx = EVP_PKEY_CTX_new_from_name(NULL, "EC", NULL);
-    if (ctx == NULL) {
-        res = SOTER_NO_MEMORY;
-        goto err;
-    }
-
-    if (!EVP_PKEY_fromdata_init(ctx)) {
-        res = SOTER_FAIL;
-        goto err;
-    }
-
-    if (!EVP_PKEY_fromdata(ctx, (EVP_PKEY**)engine_key, EVP_PKEY_KEYPAIR, params)) {
-        res = SOTER_FAIL;
-        goto err;
-    }
-
-    res = SOTER_SUCCESS;
-#endif
 
 err:
 
-#ifndef THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT
     if (d) {
         BN_clear_free(d);
     }
@@ -768,12 +441,8 @@ err:
     if (ec) {
         EC_KEY_free(ec);
     }
-#else
-    EVP_PKEY_CTX_free(ctx);
-    OSSL_PARAM_free(params);
-    OSSL_PARAM_BLD_free(bld);
-    BN_clear_free(d);
-#endif
 
     return res;
 }
+
+#endif /* THEMIS_EXPERIMENTAL_OPENSSL_3_SUPPORT */
